Extract Register from core.py
