+++ magia/factory.py
@@ -13,6 +13,7 @@ if typing.TYPE_CHECKING:
     from .constant import Constant
     from .core import Signal
     from .data_struct import OPType
+    from .register import Register
 
 
 def constant(value: int | bytes, width: int, signed: bool) -> Constant:
@@ -34,10 +35,37 @@ def create_comb_op(op_type: OPType, x: Signal, y: None | Signal | slice | int | 
     return Operation.create(op_type, x, y)
 
 
+def register(
+        width: int,
+        enable: None | Signal = None,
+        reset: None | Signal = None,
+        reset_value: None | bytes | int = None,
+        async_reset: None | Signal = None,
+        async_reset_value: None | bytes | int = None,
+        clk: None | Signal = None,
+        name: None | str = None,
+        **kwargs
+) -> Register:
+    """Create a register."""
+    return Register(
+        width=width,
+        enable=enable,
+        reset=reset,
+        reset_value=reset_value,
+        async_reset=async_reset,
+        async_reset_value=async_reset_value,
+        clk=clk,
+        name=name,
+        **kwargs
+    )
+
+
 def deferred_imports():
     from .comb_ops import Operation as OperationImported
     from .constant import Constant as ConstantImported
+    from .register import Register as RegisterImported
     globals().update({
         "Constant": ConstantImported,
         "Operation": OperationImported,
+        "Register": RegisterImported,
     })
